feat: Encrypt private key with key exchange
This would reduce the amount of data when encrypting with a list of recipients by generating a unique 128bit key that is encrypted via key exchange between the sender and the intended recipient. This would allow for a single field to be encrypted instead of an array of the same data being encrypted for each recipient.